Script name without -f and arguments without --
"Intelligently" detects it :P

Allows the script name to be the last argument after options and
then every argument after that to be options to the script. Any
script using this feature will not use --, that will just get
passed like a normal argument.

It also has "legacy" compatibility with -f, which still requires
'--' before arguments.

$@ now reflects the new behaviour.

Updated the usage to be slightly more verbose and include this
information.