Tokenize variable output on ' ' and TAB
Otherwise, subprocesses see it as all one argument. This causes
arguments to not work correctly, especially when $@ is used or
any kind of variable is substituted with a space in it.

N.B. A comment from the code:

We need to tokenize the variables we have substituted now. You might
ask, why not just tokenize everything after variable substitution?
It is because the tokenization process is rather different. For
example, if we found a #, we don't want that to be a line comment.
Mainly, we just want to seperate on spaces and tabs.